move large inline test fixtures to external files
