+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainActivity-anonymous-delegate.kt
@@ -1,0 +1,25 @@
+package com.helloworld
+
+import com.facebook.react.ReactActivity
+import com.facebook.react.ReactActivityDelegate
+import com.facebook.react.ReactRootView
+import com.swmansion.gesturehandler.react.RNGestureHandlerEnabledRootView
+
+class MainActivity : ReactActivity() {
+
+  /**
+   * Returns the name of the main component registered from JavaScript. This is used to schedule
+   * rendering of the component.
+   */
+  override fun mainComponentName: String {
+    return "HelloWorld"
+  }
+
+  override fun createReactActivityDelegate(): ReactActivityDelegate {
+    return object : ReactActivityDelegate(this, mainComponentName) {
+      override fun createRootView(): ReactRootView {
+        return RNGestureHandlerEnabledRootView(this@MainActivity)
+      }
+    }
+  }
+}
