+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainActivity-no-delegate.kt
@@ -1,0 +1,14 @@
+package com.helloworld
+
+import com.facebook.react.ReactActivity
+
+class MainActivity : ReactActivity() {
+
+  /**
+   * Returns the name of the main component registered from JavaScript. This is used to schedule
+   * rendering of the component.
+   */
+  override fun mainComponentName: String {
+    return "HelloWorld"
+  }
+}
