+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainActivity-rn064-updated.java
@@ -1,0 +1,22 @@
+package com.helloworld;
+import expo.modules.ReactActivityDelegateWrapper;
+
+import com.facebook.react.ReactActivity;
+import com.facebook.react.ReactActivityDelegate;
+
+public class MainActivity extends ReactActivity {
+
+  /**
+   * Returns the name of the main component registered from JavaScript. This is used to schedule
+   * rendering of the component.
+   */
+  @Override
+  protected String getMainComponentName() {
+    return "HelloWorld";
+  }
+
+  @Override
+  protected ReactActivityDelegate createReactActivityDelegate() {
+    return new ReactActivityDelegateWrapper(this, new ReactActivityDelegate(this, getMainComponentName()));
+  }
+}
