+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainActivity-rn064-updated.kt
@@ -1,0 +1,20 @@
+package com.helloworld
+import expo.modules.ReactActivityDelegateWrapper
+
+import com.facebook.react.ReactActivity
+import com.facebook.react.ReactActivityDelegate
+
+class MainActivity : ReactActivity() {
+
+  /**
+   * Returns the name of the main component registered from JavaScript. This is used to schedule
+   * rendering of the component.
+   */
+  override fun mainComponentName: String {
+    return "HelloWorld"
+  }
+
+  override fun createReactActivityDelegate(): ReactActivityDelegate {
+    return ReactActivityDelegateWrapper(this, ReactActivityDelegate(this, mainComponentName))
+  }
+}
