+++ packages/install-expo-modules/src/plugins/android/__tests__/fixtures/MainApplication-rn064-updated.kt
@@ -1,0 +1,44 @@
+package com.helloworld
+import expo.modules.ApplicationLifecycleDispatcher
+import expo.modules.ReactNativeHostWrapper
+
+import android.app.Application
+import android.content.res.Configuration;
+import com.facebook.react.PackageList
+import com.facebook.react.ReactApplication
+import com.facebook.react.ReactNativeHost
+import com.facebook.react.ReactPackage
+import com.facebook.soloader.SoLoader
+
+class MainApplication : Application(), ReactApplication {
+  private val mReactNativeHost: ReactNativeHost = ReactNativeHostWrapper(this, object : ReactNativeHost(this) {
+    override fun getUseDeveloperSupport(): Boolean {
+      return BuildConfig.DEBUG
+    }
+
+    override fun getPackages(): List<ReactPackage> {
+      return PackageList(this).packages
+    }
+
+    override fun getJSMainModuleName(): String {
+      return "index"
+    }
+  })
+
+  override fun getReactNativeHost(): ReactNativeHost {
+    return mReactNativeHost
+  }
+
+  override fun onCreate() {
+    super.onCreate()
+    SoLoader.init(this,  /* native exopackage */ false)
+    // comment out initializeFlipper to keep test simpler
+    // initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
+    ApplicationLifecycleDispatcher.onApplicationCreate(this)
+  }
+
+  override fun onConfigurationChanged(newConfig: Configuration) {
+    super.onConfigurationChanged(newConfig)
+    ApplicationLifecycleDispatcher.onConfigurationChanged(this, newConfig)
+  }
+}
