+++ packages/install-expo-modules/src/plugins/ios/__tests__/fixtures/AppDelegate-updated.h
@@ -1,0 +1,9 @@
+#import <React/RCTBridgeDelegate.h>
+#import <Expo/Expo.h>
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : EXAppDelegateWrapper <UIApplicationDelegate, RCTBridgeDelegate>
+
+@property (nonatomic, strong) UIWindow *window;
+
+@end
