+++ packages/install-expo-modules/src/plugins/ios/__tests__/fixtures/Podfile-with-post-integrate
@@ -1,0 +1,9 @@
+target 'HelloWorld' do
+  config = use_native_modules!
+
+  use_react_native!(:path => config["reactNativePath"])
+
+  post_integrate do |installer|
+    puts "Existing post_integrate hook"
+  end
+end
