+++ packages/install-expo-modules/src/plugins/ios/__tests__/fixtures/Podfile-with-post-integrate-updated
@@ -1,0 +1,16 @@
+require File.join(File.dirname(`node --print "require.resolve('expo/package.json')"`), "scripts/autolinking")
+target 'HelloWorld' do
+  use_expo_modules!
+  config = use_native_modules!
+
+  use_react_native!(:path => config["reactNativePath"])
+
+  post_integrate do |installer|
+    begin
+      expo_patch_react_imports!(installer)
+    rescue => e
+      Pod::UI.warn e
+    end
+    puts "Existing post_integrate hook"
+  end
+end
