+++ packages/install-expo-modules/src/plugins/ios/__tests__/withIosModulesAppDelegate-test.ts
@@ -8,28 +8,14 @@ import {
   updateModulesAppDelegateSwift,
 } from '../withIosModulesAppDelegate';
 
+const fixturesPath = path.resolve(__dirname, 'fixtures');
+
 describe(updateModulesAppDelegateObjcHeader, () => {
-  it('should migrate from classic RN AppDelegate header', () => {
-    const rawContents = `
-#import <React/RCTBridgeDelegate.h>
-#import <UIKit/UIKit.h>
-
-@interface AppDelegate : UIResponder <UIApplicationDelegate, RCTBridgeDelegate>
-
-@property (nonatomic, strong) UIWindow *window;
-
-@end`;
-
-    const expectContents = `
-#import <React/RCTBridgeDelegate.h>
-#import <Expo/Expo.h>
-#import <UIKit/UIKit.h>
-
-@interface AppDelegate : EXAppDelegateWrapper <UIApplicationDelegate, RCTBridgeDelegate>
-
-@property (nonatomic, strong) UIWindow *window;
-
-@end`;
+  it('should migrate from classic RN AppDelegate header', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'AppDelegate.h'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'AppDelegate-updated.h'), 'utf8'),
+    ]);
 
     const sdkVersion = getLatestSdkVersion().expoSdkVersion;
     const contents = updateModulesAppDelegateObjcHeader(rawContents, sdkVersion);
@@ -41,89 +27,11 @@ describe(updateModulesAppDelegateObjcHeader, () => {
 });
 
 describe(updateModulesAppDelegateObjcImpl, () => {
-  it('should migrate from classic RN AppDelegate implementation', () => {
-    const rawContents = `
-#import "AppDelegate.h"
-
-#import <React/RCTBridge.h>
-#import <React/RCTBundleURLProvider.h>
-#import <React/RCTRootView.h>
-
-@implementation AppDelegate
-
-- (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
-{
-#ifdef FB_SONARKIT_ENABLED
-  InitializeFlipper(application);
-#endif
-
-  RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:launchOptions];
-  RCTRootView *rootView = [[RCTRootView alloc] initWithBridge:bridge
-                                               moduleName:@"RN0633"
-                                               initialProperties:nil];
-
-  rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f green:1.0f blue:1.0f alpha:1];
-
-  self.window = [[UIWindow alloc] initWithFrame:[UIScreen mainScreen].bounds];
-  UIViewController *rootViewController = [UIViewController new];
-  rootViewController.view = rootView;
-  self.window.rootViewController = rootViewController;
-  [self.window makeKeyAndVisible];
-  return YES;
-}
-
-- (NSURL *)sourceURLForBridge:(RCTBridge *)bridge
-{
-#if DEBUG
-  return [[RCTBundleURLProvider sharedSettings] jsBundleURLForBundleRoot:@"index" fallbackResource:nil];
-#else
-  return [[NSBundle mainBundle] URLForResource:@"main" withExtension:@"jsbundle"];
-#endif
-}
-
-@end`;
-
-    const expectContents = `
-#import "AppDelegate.h"
-
-#import <React/RCTBridge.h>
-#import <React/RCTBundleURLProvider.h>
-#import <React/RCTRootView.h>
-
-@implementation AppDelegate
-
-- (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
-{
-#ifdef FB_SONARKIT_ENABLED
-  InitializeFlipper(application);
-#endif
-
-  RCTBridge *bridge = [self.reactDelegate createBridgeWithDelegate:self launchOptions:launchOptions];
-  RCTRootView *rootView = [self.reactDelegate createRootViewWithBridge:bridge
-                                               moduleName:@"RN0633"
-                                               initialProperties:nil];
-
-  rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f green:1.0f blue:1.0f alpha:1];
-
-  self.window = [[UIWindow alloc] initWithFrame:[UIScreen mainScreen].bounds];
-  UIViewController *rootViewController = [self.reactDelegate createRootViewController];
-  rootViewController.view = rootView;
-  self.window.rootViewController = rootViewController;
-  [self.window makeKeyAndVisible];
-  [super application:application didFinishLaunchingWithOptions:launchOptions];
-  return YES;
-}
-
-- (NSURL *)sourceURLForBridge:(RCTBridge *)bridge
-{
-#if DEBUG
-  return [[RCTBundleURLProvider sharedSettings] jsBundleURLForBundleRoot:@"index" fallbackResource:nil];
-#else
-  return [[NSBundle mainBundle] URLForResource:@"main" withExtension:@"jsbundle"];
-#endif
-}
-
-@end`;
+  it('should migrate from classic RN AppDelegate implementation', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'AppDelegate.m'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'AppDelegate-updated.m'), 'utf8'),
+    ]);
 
     const sdkVersion = getLatestSdkVersion().expoSdkVersion;
     const contents = updateModulesAppDelegateObjcImpl(rawContents, sdkVersion);
@@ -135,96 +43,23 @@ describe(updateModulesAppDelegateObjcImpl, () => {
 });
 
 describe(updateModulesAppDelegateSwift, () => {
-  it('should migrate from basic RN AppDelegate.swift', () => {
-    const contents = `
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: NSObject, UIApplicationDelegate, RCTBridgeDelegate {
-  var window: UIWindow?
-
-  func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-    let bridge = RCTBridge(delegate: self, launchOptions: launchOptions)!
-    let rootView = RCTRootView(bridge: bridge, moduleName: "HelloWorld", initialProperties: nil)
-
-    if #available(iOS 13.0, *) {
-      rootView.backgroundColor = UIColor.systemBackground
-    } else {
-      rootView.backgroundColor = UIColor.white
-    }
-
-    self.window = UIWindow(frame: UIScreen.main.bounds);
-    let rootViewController = UIViewController()
-    rootViewController.view = rootView
-    self.window?.rootViewController = rootViewController
-    self.window?.makeKeyAndVisible()
-    return true
-  }
-
-  func sourceURL(for bridge: RCTBridge!) -> URL! {
-    #if DEBUG
-    return RCTBundleURLProvider.sharedSettings()?.jsBundleURL(forBundleRoot: "index", fallbackResource: nil)
-    #else
-    return Bundle.main.url(forResource: "main", withExtension: "jsBundle")
-    #endif
-  }
-}
-`;
-
-    const expectContents = `
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: AppDelegateWrapper, RCTBridgeDelegate {
-  var window: UIWindow?
-
-  func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-    let bridge = reactDelegate.createBridge(delegate: self, launchOptions: launchOptions)!
-    let rootView = reactDelegate.createRootView(bridge: bridge, moduleName: "HelloWorld", initialProperties: nil)
-
-    if #available(iOS 13.0, *) {
-      rootView.backgroundColor = UIColor.systemBackground
-    } else {
-      rootView.backgroundColor = UIColor.white
-    }
-
-    self.window = UIWindow(frame: UIScreen.main.bounds);
-    let rootViewController = reactDelegate.createRootViewController()
-    rootViewController.view = rootView
-    self.window?.rootViewController = rootViewController
-    self.window?.makeKeyAndVisible()
-    super.application(application, didFinishLaunchingWithOptions: launchOptions)
-    return true
-  }
-
-  func sourceURL(for bridge: RCTBridge!) -> URL! {
-    #if DEBUG
-    return RCTBundleURLProvider.sharedSettings()?.jsBundleURL(forBundleRoot: "index", fallbackResource: nil)
-    #else
-    return Bundle.main.url(forResource: "main", withExtension: "jsBundle")
-    #endif
-  }
-}
-`;
+  it('should migrate from basic RN AppDelegate.swift', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'AppDelegate.swift'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'AppDelegate-updated.swift'), 'utf8'),
+    ]);
 
     const sdkVersion = getLatestSdkVersion().expoSdkVersion;
-    expect(updateModulesAppDelegateSwift(contents, sdkVersion)).toEqual(expectContents);
+    expect(updateModulesAppDelegateSwift(rawContents, sdkVersion)).toEqual(expectContents);
   });
 });
 
 describe('withIosModulesAppDelegate sdkVersion snapshots', () => {
-  const objcHeaderFixture = fs.readFileSync(
-    path.join(__dirname, 'fixtures', 'AppDelegate.h'),
-    'utf-8'
-  );
-  const objcImplFixture = fs.readFileSync(
-    path.join(__dirname, 'fixtures', 'AppDelegate.m'),
-    'utf-8'
-  );
-  const swiftFixture = fs.readFileSync(
-    path.join(__dirname, 'fixtures', 'AppDelegate.swift'),
-    'utf-8'
-  );
+  const [objcHeaderFixture, objcImplFixture, swiftFixture] = [
+    fs.readFileSync(path.join(fixturesPath, 'AppDelegate.h'), 'utf8'),
+    fs.readFileSync(path.join(fixturesPath, 'AppDelegate.m'), 'utf8'),
+    fs.readFileSync(path.join(fixturesPath, 'AppDelegate.swift'), 'utf8'),
+  ];
 
   ['43.0.0', '44.0.0'].forEach(sdkVersion => {
     it(`sdkVersion ${sdkVersion}`, () => {
